pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Paula Rabelo <paula_rabelo2010@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
pytest = "^8.3.4"
factory-boy = "^3.3.1"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
